chore: move web3 utils
