+++ scripts/arbitrum_earn.py
@@ -6,7 +6,7 @@ from config.constants import MULTICHAIN_CONFIG
 from config.enums import Network
 from src.earner import Earner
 from src.utils import get_secret
-from src.utils import get_strategies_and_vaults
+from src.web3_utils import get_strategies_and_vaults
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
