+++ scripts/earn.py
@@ -8,7 +8,7 @@ from config.enums import Network
 from src.earner import Earner
 from src.utils import get_node_url
 from src.utils import get_secret
-from src.utils import get_strategies_and_vaults
+from src.web3_utils import get_strategies_and_vaults
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
