+++ scripts/ftm_earn.py
@@ -7,7 +7,7 @@ from config.constants import MULTICHAIN_CONFIG
 from config.enums import Network
 from src.earner import Earner
 from src.utils import get_secret
-from src.utils import get_strategy_from_vault
+from src.web3_utils import get_strategy_from_vault
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
