+++ scripts/harvest_eth.py
@@ -19,7 +19,7 @@ from config.enums import Network
 from src.general_harvester import GeneralHarvester
 from src.tx_utils import get_latest_base_fee
 from src.utils import get_abi
-from src.utils import get_last_harvest_times
+from src.web3_utils import get_last_harvest_times
 from src.utils import get_secret
 from src.misc_utils import hours
 from src.misc_utils import seconds_to_blocks
