+++ scripts/poly_harvest.py
@@ -10,7 +10,7 @@ from config.constants import POLY_OLD_STRATEGY_2
 from config.enums import Network
 from src.general_harvester import GeneralHarvester
 from src.utils import get_secret
-from src.utils import get_strategies_from_registry
+from src.web3_utils import get_strategies_from_registry
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
