+++ src/eth/stability_executor.py
@@ -10,7 +10,7 @@ from config.enums import Network
 from src.tx_utils import get_effective_gas_price
 from src.tx_utils import get_gas_price_of_tx
 from src.tx_utils import get_priority_fee
-from src.utils import confirm_transaction
+from src.web3_utils import confirm_transaction
 from src.utils import get_abi
 from src.utils import send_error_to_discord
 from src.utils import send_success_to_discord
