+++ src/rebaser.py
@@ -16,7 +16,7 @@ from config.enums import Network
 from src.tx_utils import get_effective_gas_price
 from src.tx_utils import get_gas_price_of_tx
 from src.tx_utils import get_priority_fee
-from src.utils import confirm_transaction
+from src.web3_utils import confirm_transaction
 from src.utils import get_hash_from_failed_tx_error
 from src.misc_utils import hours
 from src.utils import send_rebase_error_to_discord
